requirements.txt
streamlit
pandas
langchain-groq
langchain-experimental
langchain-classic
python-dotenv
tabulate